lodview util changed in api and test
