+++ src/util/ld_util.py
@@ -1,14 +1,13 @@
-import json
 import logging
 import requests
 import validators
-
 from requests.exceptions import ConnectionError, HTTPError
-from rdflib import Graph, URIRef, Literal, BNode, Namespace
-from rdflib.namespace import RDF, RDFS, SDO, SKOS, DCTERMS
-from typing import Optional, List
+from rdflib import Graph, URIRef, Namespace
+from rdflib.namespace._RDF import RDF
+from rdflib.namespace._SDO import SDO
+from rdflib.namespace._SKOS import SKOS
+from typing import Optional
 from urllib.parse import urlparse, urlunparse
-
 from enum import Enum
 from models.DatasetApiUriLevel import DatasetApiUriLevel
 from models.ResourceApiUriLevel import ResourceApiUriLevel
@@ -394,18 +393,6 @@ def get_label_for_has_part(resource_uri: str, sparql_endpoint: str) -> Graph:
     return sparql_construct_query(sparql_endpoint, query_construct_labels_for_has_part)
 
 
-def json_ld_structured_data_for_resource(rdf_graph: Graph, resource_url: str) -> str:
-    """Returns a serialized graph in JSON_LD format with triples for the resource.
-    :param rdf_graph: Graph object containing the triples.
-    :param resource_url: the main resource for which the structured data is returned.
-    """
-    return rdf_graph.serialize(
-        format="json-ld",
-        context=dict(rdf_graph.namespaces()),
-        auto_compact=True,
-    )
-
-
 def sparql_construct_query(sparql_endpoint: str, query: str) -> Graph:
     """Sends a SPARQL CONSTRUCT query to the SPARQL endpoint and returns the result parsed into a Graph.
     raises a ConnectionError when the sparql endpoint can not be reached, or
@@ -416,233 +403,3 @@ def sparql_construct_query(sparql_endpoint: str, query: str) -> Graph:
     if resp.status_code == 200:
         g.parse(data=resp.text, format="xml")
     return g
-
-
-# ========= JSON generator functions for lod-view ==========
-
-
-def json_header_from_rdf_graph(
-    rdf_graph: Graph, resource_url: str
-) -> Optional[List[dict]]:
-    """Generates the data for the header in resource.html"""
-    json_header = []
-    try:
-        json_header = [
-            {
-                "title": [
-                    str(label)
-                    for label in rdf_graph.objects(URIRef(resource_url), SKOS.prefLabel)
-                    if Literal(label).language == "nl"
-                ]
-                + [
-                    str(name)
-                    for name in rdf_graph.objects(URIRef(resource_url), SDO.name)
-                ]
-                + [
-                    str(name)
-                    for name in rdf_graph.objects(URIRef(resource_url), DCTERMS.title)
-                ],
-                "o": {
-                    "uri": str(o),
-                    "prefix": rdf_graph.compute_qname(str(o))[0],
-                    "namespace": str(rdf_graph.compute_qname(str(o))[1]),
-                    "property": rdf_graph.compute_qname(str(o))[2],
-                },
-            }
-            for o in rdf_graph.objects(
-                subject=URIRef(resource_url), predicate=URIRef(RDF.type)
-            )
-            if str(rdf_graph.compute_qname(str(o))[1]) in (str(SDO), str(SKOS))
-        ]
-        return json_header
-    except Exception as e:
-        logger.exception(f"Error in json_header_from_rdf_graph: {str(e)}")
-        logger.error(json_header)
-    return json_header
-
-
-def json_iri_iri_from_rdf_graph(
-    rdf_graph: Graph, resource_url: str
-) -> Optional[List[dict]]:
-    """Generates part of the LOD view data for resource.html that with: (<uri> <uri> <uri>)"""
-    json_iri_iri = []
-    try:
-        json_iri_iri = [
-            {
-                "p": {
-                    "uri": str(p),
-                    "prefix": rdf_graph.compute_qname(str(p))[0],
-                    "namespace": str(rdf_graph.compute_qname(str(p))[1]),
-                    "property": rdf_graph.compute_qname(str(p))[2],
-                },
-                "o": {
-                    "uri": str(o),
-                    "literal_form": [
-                        f"{str(lf)} @{Literal(lf).language}"
-                        for lf in rdf_graph.objects(
-                            o, URIRef(f"{SKOSXL_NS}literalForm")
-                        )
-                    ],
-                    "prefix": rdf_graph.compute_qname(str(o))[0],
-                    "namespace": str(rdf_graph.compute_qname(str(o))[1]),
-                    "property": rdf_graph.compute_qname(str(o))[2],
-                    "pref_label": [
-                        str(label)
-                        for label in rdf_graph.objects(o, SKOS.prefLabel)
-                        if Literal(label).language == "nl"
-                    ],
-                    "parent_label": [
-                        str(label) for label in rdf_graph.objects(o, SDO.name)
-                    ],
-                    "part_label": [
-                        str(label) for label in rdf_graph.objects(o, SDO.name)
-                    ],
-                },
-            }
-            for (p, o) in rdf_graph.predicate_objects(subject=URIRef(resource_url))
-            if p != RDF.type and isinstance(o, URIRef) and not str(o).endswith("/")
-        ]
-        # URIs that end with a '/' can not be split by rdflib. Therefore we add them separately.
-        for p, o in rdf_graph.predicate_objects(subject=URIRef(resource_url)):
-            if isinstance(o, URIRef) and str(o).endswith("/"):
-                json_iri_iri += [
-                    {
-                        "p": {
-                            "uri": str(p),
-                            "prefix": rdf_graph.compute_qname(str(p))[0],
-                            "namespace": str(rdf_graph.compute_qname(str(p))[1]),
-                            "property": rdf_graph.compute_qname(str(p))[2],
-                        },
-                        "o": {
-                            "uri": str(o),
-                            "literal_form": [],
-                            "prefix": "",
-                            "namespace": "",
-                            "property": "",
-                            "pref_label": [],
-                        },
-                    }
-                ]
-        return json_iri_iri
-    except Exception as e:
-        logger.exception(f"Error in json_iri_iri_from_rdf_graph: {str(e)}")
-        logger.error(json_iri_iri)
-    return json_iri_iri
-
-
-def json_iri_lit_from_rdf_graph(
-    rdf_graph: Graph, resource_url: str
-) -> Optional[List[dict]]:
-    """Generates part of the LOD view data for resource.html (<uri> <uri> literal)"""
-    json_iri_lit = []
-    try:
-        json_iri_lit = [
-            {
-                "p": {
-                    "uri": str(p),
-                    "prefix": rdf_graph.compute_qname(str(p))[0],
-                    "namespace": str(rdf_graph.compute_qname(str(p))[1]),
-                    "property": rdf_graph.compute_qname(str(p))[2],
-                },
-                "o": {
-                    "literal_value": (
-                        f"{str(o)} @{o.language}" if o.language else f"{str(o)}"
-                    ),
-                    "datatype": str(o.datatype) if o.datatype is not None else "",
-                    "datatype_prefix": (
-                        rdf_graph.compute_qname(str(o.datatype))[0]
-                        if o.datatype is not None
-                        else ""
-                    ),
-                    "datatype_namespace": (
-                        str(rdf_graph.compute_qname(str(o.datatype))[1])
-                        if o.datatype is not None
-                        else ""
-                    ),
-                    "datatype_property": (
-                        rdf_graph.compute_qname(str(o.datatype))[2]
-                        if o.datatype is not None
-                        else ""
-                    ),
-                },
-            }
-            for p, o in rdf_graph.predicate_objects(subject=URIRef(resource_url))
-            if isinstance(o, Literal)
-        ]
-        return json_iri_lit
-    except Exception as e:
-        logger.exception(f"Error in json_iri_lit_from_rdf_graph: {str(e)}")
-        logger.error(json_iri_lit)
-    return json_iri_lit
-
-
-def json_iri_bnode_from_rdf_graph(
-    rdf_graph: Graph, resource_url: str
-) -> Optional[List[dict]]:
-    """Generates part of the LOD view data for resource.html matching: (<uri> <uri> [bnode])"""
-    # TODO: make sure that we also handle bnodes that have a type and a property with a value.
-    json_iri_bnode = []
-    try:
-        for p, o in rdf_graph.predicate_objects(subject=URIRef(resource_url)):
-            if p != RDF.type and isinstance(o, BNode):
-                bnode_content = [
-                    {
-                        "pred": {
-                            "uri": str(bnode_pred),
-                            "prefix": rdf_graph.compute_qname(str(bnode_pred))[0],
-                            "namespace": str(
-                                rdf_graph.compute_qname(str(bnode_pred))[1]
-                            ),
-                            "property": rdf_graph.compute_qname(str(bnode_pred))[2],
-                        },
-                        "obj": (
-                            {
-                                "uri": str(bnode_obj),
-                                "prefix": rdf_graph.compute_qname(str(bnode_obj))[0],
-                                "namespace": str(
-                                    rdf_graph.compute_qname(str(bnode_obj))[1]
-                                ),
-                                "property": rdf_graph.compute_qname(str(bnode_obj))[2],
-                                "pref_label": [
-                                    f"{str(pl)} @{Literal(pl).language}"
-                                    for pl in rdf_graph.objects(
-                                        bnode_obj, SKOS.prefLabel
-                                    )
-                                ],
-                            }
-                            if isinstance(bnode_obj, URIRef)
-                            else (
-                                {
-                                    "label": (
-                                        f"{str(bnode_obj)} @{bnode_obj.language}"
-                                        if bnode_obj.language
-                                        else f"{str(bnode_obj)}"
-                                    )
-                                }
-                                if isinstance(bnode_obj, Literal)
-                                else str(bnode_obj)
-                            )
-                        ),
-                    }
-                    for (bnode_pred, bnode_obj) in rdf_graph.predicate_objects(
-                        subject=o
-                    )
-                    if bnode_obj
-                    != RDFS.Resource  # and bnode_pred.property != "associatedMedia"
-                ]
-                json_iri_bnode.append(
-                    {
-                        "p": {
-                            "uri": str(p),
-                            "prefix": rdf_graph.compute_qname(str(p))[0],
-                            "namespace": str(rdf_graph.compute_qname(str(p))[1]),
-                            "property": rdf_graph.compute_qname(str(p))[2],
-                        },
-                        "o": bnode_content,
-                    }
-                )
-        return json_iri_bnode
-    except Exception as e:
-        logger.exception(f"Error in json_iri_bnode_from_rdf_graph: {str(e)}")
-        logger.error(json.dumps(json_iri_bnode, indent=4))
-    return json_iri_bnode
